crop-backend/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
joblib==1.3.2
pydantic==2.5.0
google-earth-engine-api==0.1.384
requests==2.31.0
geopandas==0.14.0
pymongo==4.6.0
python-dotenv==1.0.0
loguru==0.7.2
kaggle==1.5.16
python-multipart==0.0.6